+++ aes/src/backends.rs
@@ -1,24 +1,13 @@
 pub(crate) mod soft;
 
-#[cfg(all(target_arch = "aarch64", not(aes_backend = "soft")))]
-pub(crate) mod aarch64_aes;
-
-#[cfg(all(
-    any(target_arch = "x86_64", target_arch = "x86"),
-    not(aes_backend = "soft")
-))]
-pub(crate) mod x86_aes;
-
-#[cfg(all(
-    any(aes_backend = "avx256", aes_backend = "avx512"),
-    any(target_arch = "x86_64", target_arch = "x86"),
-    not(aes_backend = "soft"),
-))]
-pub(crate) mod x86_vaes256;
-
-#[cfg(all(
-    aes_backend = "avx512",
-    any(target_arch = "x86_64", target_arch = "x86"),
-    not(aes_backend = "soft"),
-))]
-pub(crate) mod x86_vaes512;
+cpubits::cfg_if! {
+    if #[cfg(aes_backend = "soft")] {
+        // Do not add other backends if software backend is forced
+    } else if #[cfg(target_arch = "aarch64")] {
+        pub(crate) mod aarch64_aes;
+    } else if #[cfg(any(target_arch = "x86_64", target_arch = "x86"))] {
+        pub(crate) mod x86_aes;
+        pub(crate) mod x86_vaes256;
+        pub(crate) mod x86_vaes512;
+    }
+}