+++ aes/src/backends/aarch64_aes.rs
@@ -7,6 +7,7 @@ use cipher::{
 
 mod encdec;
 mod expand;
+mod utils;
 
 #[cfg(feature = "hazmat")]
 pub(crate) mod hazmat;
@@ -36,15 +37,13 @@ pub(crate) struct Aes<const RK: usize> {
 impl<const RK: usize> Aes<RK> {
     #[inline]
     #[target_feature(enable = "aes")]
-    // TODO(MSRV-1.86): remove `unsafe`
-    pub(crate) unsafe fn encrypt(&self, f: impl BlockCipherEncClosure<BlockSize = U16>) {
+    pub(crate) fn encrypt(&self, f: impl BlockCipherEncClosure<BlockSize = U16>) {
         f.call(self);
     }
 
     #[inline]
     #[target_feature(enable = "aes")]
-    // TODO(MSRV-1.86): remove `unsafe`
-    pub(crate) unsafe fn decrypt(&self, f: impl BlockCipherDecClosure<BlockSize = U16>) {
+    pub(crate) fn decrypt(&self, f: impl BlockCipherDecClosure<BlockSize = U16>) {
         f.call(self);
     }
 }
@@ -69,7 +68,7 @@ impl<const RK: usize> BlockCipherEncBackend for Aes<RK> {
     fn encrypt_par_blocks(&self, blocks: InOut<'_, '_, ParBlocks<Self>>) {
         // SAFETY: this trait impl is used only by the `Self::encrypt` method marked with
         // `#[target_feature(enable = "aes")]`
-        unsafe { encdec::encrypt_par(&self.enc_rk, blocks) };
+        unsafe { encdec::batch_encrypt(&self.enc_rk, blocks) };
     }
 }
 
@@ -85,7 +84,7 @@ impl<const RK: usize> BlockCipherDecBackend for Aes<RK> {
     fn decrypt_par_blocks(&self, blocks: InOut<'_, '_, ParBlocks<Self>>) {
         // SAFETY: this trait impl is used only by the `Self::decrypt` method marked with
         // `#[target_feature(enable = "aes")]`
-        unsafe { encdec::decrypt_par(&self.dec_rk, blocks) };
+        unsafe { encdec::batch_decrypt(&self.dec_rk, blocks) };
     }
 }
 
@@ -97,25 +96,22 @@ pub(crate) struct AesEnc<const RK: usize> {
 impl<const RK: usize> AesEnc<RK> {
     #[inline]
     #[target_feature(enable = "aes")]
-    // TODO(MSRV-1.86): remove `unsafe`
-    pub(crate) unsafe fn as_encdec(&self) -> Aes<RK> {
+    pub(crate) fn as_encdec(&self) -> Aes<RK> {
         let enc_rk = self.enc_rk;
-        let dec_rk = unsafe { expand::inv_expanded_keys(&enc_rk) };
+        let dec_rk = expand::inv_expanded_keys(&enc_rk);
         Aes { enc_rk, dec_rk }
     }
 
     #[inline]
     #[target_feature(enable = "aes")]
-    // TODO(MSRV-1.86): remove `unsafe`
-    pub(crate) unsafe fn as_dec(&self) -> AesDec<RK> {
-        let dec_rk = unsafe { expand::inv_expanded_keys(&self.enc_rk) };
+    pub(crate) fn as_dec(&self) -> AesDec<RK> {
+        let dec_rk = expand::inv_expanded_keys(&self.enc_rk);
         AesDec { dec_rk }
     }
 
     #[inline]
     #[target_feature(enable = "aes")]
-    // TODO(MSRV-1.86): remove `unsafe`
-    pub(crate) unsafe fn encrypt(&self, f: impl BlockCipherEncClosure<BlockSize = U16>) {
+    pub(crate) fn encrypt(&self, f: impl BlockCipherEncClosure<BlockSize = U16>) {
         f.call(self)
     }
 }
@@ -140,7 +136,7 @@ impl<const RK: usize> BlockCipherEncBackend for AesEnc<RK> {
     fn encrypt_par_blocks(&self, blocks: InOut<'_, '_, ParBlocks<Self>>) {
         // SAFETY: this trait impl is used only by the `Self::encrypt` method marked with
         // `#[target_feature(enable = "aes")]`
-        unsafe { encdec::encrypt_par(&self.enc_rk, blocks) };
+        unsafe { encdec::batch_encrypt(&self.enc_rk, blocks) };
     }
 }
 
@@ -152,8 +148,7 @@ pub(crate) struct AesDec<const RK: usize> {
 impl<const RK: usize> AesDec<RK> {
     #[inline]
     #[target_feature(enable = "aes")]
-    // TODO(MSRV-1.86): remove `unsafe`
-    pub(crate) unsafe fn decrypt(&self, f: impl BlockCipherDecClosure<BlockSize = U16>) {
+    pub(crate) fn decrypt(&self, f: impl BlockCipherDecClosure<BlockSize = U16>) {
         f.call(self);
     }
 }
@@ -178,7 +173,7 @@ impl<const RK: usize> BlockCipherDecBackend for AesDec<RK> {
     fn decrypt_par_blocks(&self, blocks: InOut<'_, '_, ParBlocks<Self>>) {
         // SAFETY: this trait impl is used only by the `Self::decrypt` method marked with
         // `#[target_feature(enable = "aes")]`
-        unsafe { encdec::decrypt_par(&self.dec_rk, blocks) };
+        unsafe { encdec::batch_decrypt(&self.dec_rk, blocks) };
     }
 }
 
@@ -187,10 +182,9 @@ macro_rules! impl_key_init {
         impl $name {
             #[inline]
             #[target_feature(enable = "aes")]
-            // TODO(MSRV-1.86): remove `unsafe`
-            pub(crate) unsafe fn new(key: &[u8; $key_size]) -> Self {
-                let enc_rk = unsafe { expand::expand_key(key) };
-                let dec_rk = unsafe { expand::inv_expanded_keys(&enc_rk) };
+            pub(crate) fn new(key: &[u8; $key_size]) -> Self {
+                let enc_rk = expand::expand_key(key);
+                let dec_rk = expand::inv_expanded_keys(&enc_rk);
                 Self { enc_rk, dec_rk }
             }
         }
@@ -198,9 +192,8 @@ macro_rules! impl_key_init {
         impl $name_enc {
             #[inline]
             #[target_feature(enable = "aes")]
-            // TODO(MSRV-1.86): remove `unsafe`
-            pub(crate) unsafe fn new(key: &[u8; $key_size]) -> Self {
-                let enc_rk = unsafe { expand::expand_key(key) };
+            pub(crate) fn new(key: &[u8; $key_size]) -> Self {
+                let enc_rk = expand::expand_key(key);
                 Self { enc_rk }
             }
         }
@@ -208,10 +201,9 @@ macro_rules! impl_key_init {
         impl $name_dec {
             #[inline]
             #[target_feature(enable = "aes")]
-            // TODO(MSRV-1.86): remove `unsafe`
-            pub(crate) unsafe fn new(key: &[u8; $key_size]) -> Self {
-                let enc_rk = unsafe { expand::expand_key(key) };
-                let dec_rk = unsafe { expand::inv_expanded_keys(&enc_rk) };
+            pub(crate) fn new(key: &[u8; $key_size]) -> Self {
+                let enc_rk = expand::expand_key(key);
+                let dec_rk = expand::inv_expanded_keys(&enc_rk);
                 Self { dec_rk }
             }
         }