+++ aes/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.9.3 (UNRELEASED)
+### Changed
+- MSRV bumped to 1.89 ([#580])
+- VAES-256 and VAES-512 backends on x86 targets are now supported by default ([#580])
+
+### Removed
+- `aes_backend = "avx256"` and `aes_backend = "avx512"` configuration flags ([#580])
+
+[#580]: https://github.com/RustCrypto/block-ciphers/pull/580
+
 ## 0.9.2 (2026-07-27)
 ### Added
 - `hardware_accelerated` function ([#579])